requirements.txt
asgiref==3.5.2
atomicwrites==1.4.1
attrs==21.4.0
black==22.6.0
click==8.1.3
colorama==0.4.5
Django==4.0.6
django-cors-headers==3.13.0
djangorestframework==3.13.1
Faker==13.15.0
fire==0.4.0
gunicorn==20.1.0
importlib-metadata==4.12.0
iniconfig==1.1.1
Mako==1.2.1
Markdown==3.4.1
MarkupSafe==2.1.1
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
platformdirs==2.5.2
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.9
pytest==7.1.2
pytest-django==4.5.2
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
termcolor==1.1.0
tomli==2.0.1
typing_extensions==4.3.0
tzdata==2022.1
zipp==3.8.1
